Made required changes to the admin folders
