+++ src/App.js
@@ -9,7 +9,7 @@ import { getDocs, collection, query, where } from "firebase/firestore";
 import { db } from "./component/firebase";
 
 import Home from "./component/Home";
-import UserList from "./component/Admin/UserList";
+import UserList from "./component/admin_test/UserList.js";
 import NavBar from "./component/Navbar";
 import Footer from "./component/Footer";
 import Login from "./component/Login";
@@ -56,8 +56,8 @@ import EmailUpdateConfirmation from "./component/UserProfile/ProfileSettings/Ema
 import DeleteAccConfirmation from "./component/UserProfile/ProfileSettings/DeleteAccConfirmation";
 import UpdateProfile from "./component/UserProfile/ProfileSettings/UpdateProfile";
 
-import UserDetails from "./component/Admin/UserDetails";
-import CreateOutreachAdmin from "./component/Admin/CreateOutreachAdmin";
+//import UserDetails from "./component/admin_test/UserDetails.js";
+import CreateOutreachAdmin from "./component/admin_test/CreateOutreachAdmin.js";
 
 import TestUser from "./component/Test/Test";
 import ListUser from "./component/Test/ListUser";
@@ -148,8 +148,8 @@ function App() {
           />
           <Route path="/allnews" element={<Newscard />} />
           <Route path="/allnews/:id" element={<Readmorenews />} />
-          <Route path="/user/:uid" element={<UserDetails />} /> // Route for
-          user details
+          {/* <Route path="/user/:uid" element={<UserDetails />} /> // Route for
+          user details */}
           <Route
             path="/verifyemail"
             element={
